AIadded (not complete), Minor tweaks)
AI lacks attacking logic, due to lack of attacking mechanism, little changes made to keep everything together. New objects added to "Enemies" object on each level. From now on for each enemy, there has to be an enemy starting patrol position, and it has to be assigned in the inspector